backend/nature_go/requirements.txt
django
django-cors-headers
django-filter
djangorestframework
drf-extra-fields
numpy
requests
Pillow
gunicorn
